+++ src/config/mod.rs
@@ -1,6 +1,3 @@
 mod schema;
 
-pub use schema::{
-    Config, ConfigDecoder, ConfigError, ConfigPath, ConfigSchema, JsonConfigDecoder,
-    TomlConfigDecoder, YamlConfigDecoder,
-};
+pub use schema::{Config, ConfigError, ConfigPath};